08_user board 1:n model relation

+++ boards/views.py
@@ -28,7 +28,12 @@ def create(request):
             # # board = Board(title=title, content=content)
             # # board.save()
             # board = Board.objects.create(title=title, content=content)
-            board = form.save()
+            # board = form.save()
+            # board를 바로 저장하지 않고, 현재 user를 넣고 저장
+            # 실제 DB에 반영 전까지의 단계를 진행하고, 그 중간에 user정보를 request.user에서 가져온 후에 저장한다.
+            board = form.save(commit=False)
+            board.user = request.user
+            board.save()
             return redirect('boards:detail', board.pk)
     # GET 요청(혹은 다른 method)이면 기본 FORM을 생성한다.
     else:
@@ -48,29 +53,35 @@ def detail(request, board_pk):
     
 def delete(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        board.delete()
-        return redirect('boards:index')
+    if board.user == request.user:
+        if request.method == 'POST':
+            board.delete()
+            return redirect('boards:index')
+        else:
+            return redirect('boards:detail', board.pk)
     else:
-        return redirect('boards:detail', board.pk)
-
+        return redirect('boards:index')
+        
 @login_required
 def update(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        form = BoardForm(request.POST, instance=board)  # 1
-        if form.is_valid():
-            # board.title = form.cleaned_data.get('title')
-            # board.content = form.cleaned_data.get('content')
-            # board.save()
-            board = form.save()                         # 2
-            return redirect('boards:detail', board.pk)
-    # GET 요청이면(수정하기 버튼을 눌렀을 때)
+    if board.user == request.user:
+        if request.method == 'POST':
+            form = BoardForm(request.POST, instance=board)  # 1
+            if form.is_valid():
+                # board.title = form.cleaned_data.get('title')
+                # board.content = form.cleaned_data.get('content')
+                # board.save()
+                board = form.save()                         # 2
+                return redirect('boards:detail', board.pk)
+        # GET 요청이면(수정하기 버튼을 눌렀을 때)
+        else:
+            form = BoardForm(instance=board)                # 3                    
+            # BoardForm을 초기화(사용자 입력 값을 넣어준 상태로)
+            # # form = BoardForm(initail={'title': board.title, 'content': board.content})
+            # form = BoardForm(initial=board.__dict__)
     else:
-        form = BoardForm(instance=board)                # 3                    
-        # BoardForm을 초기화(사용자 입력 값을 넣어준 상태로)
-        # # form = BoardForm(initail={'title': board.title, 'content': board.content})
-        # form = BoardForm(initial=board.__dict__)
+        return redirect('boards:index')
     # 1. POST : 요청에서 검증에 실패하였을때, 오류 메세지가 포함된 상태
     # 2. GET : 요청에서 초기화된 상태
     context = {
